# apps/web/public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MoonPlants API",
    "version": "1.0.0",
    "description": "REST API for MoonPlants — smart plant monitoring and automated watering system. Consists of two namespaces:\n\n- **`/api/v1/`** — User-facing endpoints (Supabase JWT auth)\n- **`/api/iot/v1/`** — Device-facing endpoints (HMAC-SHA256 auth)",
    "contact": {
      "name": "MoonPlants"
    },
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    },
    {
      "url": "https://your-production-domain.com",
      "description": "Production (replace with actual domain)"
    }
  ],
...
  "components": {
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string",
            "example": "Unauthorized"
          }
        }
      },
...
      "IotMeasurementPayload": {
        "type": "object",
        "required": [
          "measuredAt",
          "soil"
        ],
        "properties": {
          "measuredAt": {
            "type": "integer",
            "description": "Unix epoch seconds",
            "example": 1709481600
          },
          "air": {
            "type": "object",
            "properties": {
              "tempC": {
                "type": "number",
                "minimum": -50,
                "maximum": 100,
                "example": 22.5
              },
              "humidityPct": {
                "type": "number",
                "minimum": 0,
                "maximum": 100,
                "example": 55
              }
            }
          },
          "lightLux": {
            "type": "number",
            "minimum": 0,
            "example": 2400
          },
          "soil": {
            "type": "array",
            "minItems": 1,
            "maxItems": 16,
            "items": {
              "type": "object",
              "required": [
                "channel",
                "moistureRaw"
              ],
              "properties": {
                "channel": {
                  "type": "integer",
                  "minimum": 1,
                  "maximum": 16
                },
                "moistureRaw": {
                  "type": "integer",
                  "minimum": 0,
                  "maximum": 65535
                }
              }
            }
          },
          "batteryV": {
            "type": "number",
            "minimum": 0,
            "maximum": 10,
            "example": 3.8
          },
          "rssiDbm": {
            "type": "integer",
            "minimum": -120,
            "maximum": 0,
            "example": -67
          }
        }
      },
      "CommandAckPayload": {
        "type": "object",
        "required": [
          "status"
        ],
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "ok",
              "failed",
              "partial"
            ]
          },
          "executedAt": {
            "type": "integer",
            "description": "Unix epoch seconds when executed"
          },
          "result": {
            "type": "object",
            "additionalProperties": true,
            "description": "Optional extra result data"
          }
        }
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Supabase JWT access token. Pass via Authorization: Bearer <token>"
      },
      "IotHmac": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Device-Signature",
        "description": "HMAC-SHA256 device signature. IoT endpoints require all 5 headers: X-Device-Id, X-Device-Seq, X-Device-Timestamp, X-Content-SHA256, X-Device-Signature. Canonical string = METHOD\\nPATH\\nX-Device-Id\\nX-Device-Seq\\nX-Device-Timestamp\\nX-Content-SHA256"
      }
    }
  },
  "paths": {
...
    "/api/iot/v1/measurements": {
      "post": {
        "tags": [
          "IoT Device"
        ],
        "summary": "Submit sensor measurements",
        "description": "ESP32 posts sensor data for all active channels. The server maps channels to plant_ids and stores measurements. Uses upsert on (device_id, seq, plant_id) for idempotency.",
        "security": [
          {
            "IotHmac": []
          }
        ],
        "parameters": [
          {
            "name": "X-Device-Id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Device UUID"
          },
          {
            "name": "X-Device-Seq",
            "in": "header",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Monotonic sequence counter stored in flash"
          },
          {
            "name": "X-Device-Timestamp",
            "in": "header",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Current Unix epoch seconds (must be within ±120s of server time)"
          },
          {
            "name": "X-Content-SHA256",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Hex SHA-256 of the request body (empty string for GET requests)"
          },
          {
            "name": "X-Device-Signature",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "base64url(HMAC-SHA256(secret, canonical_string)) where canonical_string = METHOD\\nPATH\\nX-Device-Id\\nX-Device-Seq\\nX-Device-Timestamp\\nX-Content-SHA256"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IotMeasurementPayload"
              },
              "examples": {
                "twoChannels": {
                  "summary": "2-channel measurement with all sensors",
                  "value": {
                    "measuredAt": 1709481600,
                    "air": {
                      "tempC": 22.5,
                      "humidityPct": 55
                    },
                    "lightLux": 2400,
                    "soil": [
                      {
                        "channel": 1,
                        "moistureRaw": 2450
                      },
                      {
                        "channel": 2,
                        "moistureRaw": 3100
                      }
                    ],
                    "batteryV": 3.8,
                    "rssiDbm": -67
                  }
                },
                "minimalPayload": {
                  "summary": "Minimal payload (soil only)",
                  "value": {
                    "measuredAt": 1709481600,
                    "soil": [
                      {
                        "channel": 1,
                        "moistureRaw": 2450
                      }
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Measurements ingested",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ingested": {
                      "type": "integer",
                      "description": "Number of rows stored"
                    },
                    "ignoredChannels": {
                      "type": "array",
                      "items": {
                        "type": "integer"
                      },
                      "description": "Channels with no linked plant"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid JSON or payload schema",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "HMAC auth failed or missing headers",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/iot/v1/commands/{commandId}/ack": {
      "post": {
        "tags": [
          "IoT Device"
        ],
        "summary": "Acknowledge command execution",
        "description": "Device reports whether a command succeeded or failed. On successful PUMP_WATER ack, a watering_event is auto-created.",
        "security": [
          {
            "IotHmac": []
          }
        ],
        "parameters": [
          {
            "name": "X-Device-Id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Device UUID"
          },
          {
            "name": "X-Device-Seq",
            "in": "header",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Monotonic sequence counter stored in flash"
          },
          {
            "name": "X-Device-Timestamp",
            "in": "header",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Current Unix epoch seconds (must be within ±120s of server time)"
          },
          {
            "name": "X-Content-SHA256",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Hex SHA-256 of the request body (empty string for GET requests)"
          },
          {
            "name": "X-Device-Signature",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "base64url(HMAC-SHA256(secret, canonical_string)) where canonical_string = METHOD\\nPATH\\nX-Device-Id\\nX-Device-Seq\\nX-Device-Timestamp\\nX-Content-SHA256"
          },
          {
            "name": "commandId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommandAckPayload"
              },
              "examples": {
                "success": {
                  "summary": "Command succeeded",
                  "value": {
                    "status": "ok",
                    "executedAt": 1709481700
                  }
                },
                "failed": {
                  "summary": "Command failed",
                  "value": {
                    "status": "failed",
                    "executedAt": 1709481700,
                    "result": {
                      "error": "pump_jam"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Ack received",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    },
                    "alreadyAcked": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid payload",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "HMAC auth failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Command belongs to different device",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Command not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  }
}